// questionnaire/src/protos/gen/questionnaire/CreateQuestionnaire.ts
// Original file: src/protos/questionnaire.proto


export interface CreateQuestionnaire {
  'access_token'?: (string);
  'questions'?: (string)[];
  'category'?: (string);
  'status'?: (string);
}

export interface CreateQuestionnaire__Output {
  'access_token': (string);
  'questions': (string)[];
  'category': (string);
  'status': (string);
}