// questionnaire/src/protos/gen/questionnaire/UpdateQuestionnaire.ts
// Original file: src/protos/questionnaire.proto


export interface UpdateQuestionnaire {
  'access_token'?: (string);
  'questions'?: (string)[];
  'category'?: (string);
  'status'?: (string);
  'id'?: (string);
}

export interface UpdateQuestionnaire__Output {
  'access_token': (string);
  'questions': (string)[];
  'category': (string);
  'status': (string);
  'id': (string);
}